+++ Runtime/Classes.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using DateTime = System.DateTime;
+using SimpleJSON;
 
 namespace WattanaGaming.GameJoltAPI
 {
@@ -13,6 +14,16 @@ namespace WattanaGaming.GameJoltAPI
         public string description;
         public string imageURL;
         public string achieved;
+
+        public TrophyData(JSONNode JSONData)
+        {
+            id = JSONData["id"];
+            title = JSONData["title"];
+            difficulty = JSONData["difficulty"];
+            description = JSONData["description"];
+            imageURL = JSONData["image_url"];
+            achieved = JSONData["achieved"];
+        }
 
         public override string ToString()
         {